Update repnode.sh
reparando el método de instalación de repositorio node 
aseguramos que exista /etc/apt/sources.list.d/nodesource.list 
si existe confirmamos que se encuentre el repositorio con  grep -qi 'deb.nodesource.com'  /etc/apt/sources.list.d/nodesource.list 
si los dos son verdaderos omite la instalación, si alguno da negativo o falso realiza la instalación

+++ script/repnode.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 set -o pipefail
 ##repositorios para node js y npm
-echo
-echo "instalando repo nodejs"
-curl -sL https://deb.nodesource.com/setup_11.x | bash -
-apt-get install -y nodejs
+if [ -f /etc/apt/sources.list.d/nodesource.list ] && grep -qi 'deb.nodesource.com' /etc/apt/sources.list.d/nodesource.list; then
+  echo "repositorio nodejs instalado"
+else 
+  echo
+  echo "instalando repo nodejs"
+  curl -sL https://deb.nodesource.com/setup_11.x | bash -
+  apt-get install -y nodejs
+fi
